pyproject.toml
[project]
name = "p1-seafoam-cicada"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "ipykernel>=7.1.0",
    "langchain>=1.2.3",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.6",
    "langgraph-cli[inmem]>=0.4.11",
    "langserve>=0.3.3",
    "langsmith>=0.6.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "sse-starlette<3.0.0",
    "uvicorn>=0.40.0",
]
